avoid use after free in OpcTcpMessages
To avoid use of already deleted instances of OpcTcpMessages we make
it a shared_ptr instance itself, give outputChannel as a shared_ptr
instead of a reference and store outputChannel as a weak_ptr. We can now
make sure our OpcTcpMessages instance is still valid when we use it in
an async operation and check if the parent OpcTcpConnection (aka
OutputChannel) is still alive.

This patch is much more intrusive then the previously suggested hack
which simply deleted the subscription callback but it should be
much less fragile.