limit sample rate for subscriptions
For values with a higher update rate then subscription publishing
interval InternalSubscription queued as many TriggeredDataChange events
as came in. For a value with an update rate of 1kHz in a subscription
with a publishing rate of 1Hz it queued 1000 events and - even if the
client did register a queue size of 1 - sent them to the client.

This patch forces a queue size of 1.