allow multiple instances when using classic debug interface
When intruducing spdlog I tried to keep the external interface
compatible with the old style debug approach.

For multiple instances of clients or servers this failed, as spdlog
keeps track of the names of instanciated loggers and denies to create
duplicate logger names. To allow to use the classic interface with
multiple instances, check if there already is a logger with the name
we want to create.