Fix handling of ArrayDimensions in Variant
With the current implementation, reading multi-dimensional values
does not work as expected, the ArrayDimensions information is missing.
This is caused by the Dimensions vector being ignored when creating
a Variant from another Variant.

This patch fixes the issue by copying the Dimensions vector in
Variant(const Variant &) and operator=(const Variant &).